src/app/extensions/organization-hierarchies/shared/hierarchy-path/hierarchy-path.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { NgbTooltipModule } from '@ng-bootstrap/ng-bootstrap';

import { HierarchyPathComponent } from './hierarchy-path.component';

describe('Hierarchy Path Component', () => {
  let component: HierarchyPathComponent;
  let fixture: ComponentFixture<HierarchyPathComponent>;
  let element: HTMLElement;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      declarations: [HierarchyPathComponent],
      imports: [NgbTooltipModule],
    }).compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(HierarchyPathComponent);
    component = fixture.componentInstance;
    element = fixture.nativeElement;
  });

  it('should be created', () => {
    expect(component).toBeTruthy();
    expect(element).toBeTruthy();
    expect(() => fixture.detectChanges()).not.toThrow();
  });
});
